# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/profile/create": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Create a profile",
                "description": "Endpoint to create a new user profile.",
                "operationId": "createProfile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "birthday",
                                    "place",
                                    "bio"
                                ],
                                "properties": {
                                    "id": {
                                        "description": "ID of the profile",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "birthday": {
                                        "description": "Birth date for the user",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2006-06-06"
                                    },
                                    "place": {
                                        "description": "Place address of user",
                                        "type": "string",
                                        "example": "Pursat province"
                                    },
                                    "bio": {
                                        "description": "Biography or description of the user",
                                        "type": "string",
                                        "example": "Passionate about coding and technology"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Profile created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profile created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid input"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/profile/show/{id}": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Show a profile",
                "description": "Endpoint to fetch details of a user profile by ID.",
                "operationId": "showProfile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the profile to fetch",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile details fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "birthday": {
                                            "description": "Birth date for the user",
                                            "type": "string",
                                            "format": "date",
                                            "example": "2006-06-06"
                                        },
                                        "place": {
                                            "description": "Place address of user",
                                            "type": "string",
                                            "example": "Pursat province"
                                        },
                                        "bio": {
                                            "description": "Biography or description of the user",
                                            "type": "string",
                                            "example": "Passionate about coding and technology"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Profile not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Profile not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/profile/update/{id}": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update a profile",
                "description": "Endpoint to update an existing user profile.",
                "operationId": "updateProfile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the profile to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "birthday": {
                                        "description": "New birth date for the user",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2006-06-06"
                                    },
                                    "place": {
                                        "description": "New place address of user",
                                        "type": "string",
                                        "example": "Pursat province"
                                    },
                                    "bio": {
                                        "description": "New biography or description of the user",
                                        "type": "string",
                                        "example": "Updated biography about coding and technology"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profile updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid input"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Profile not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Profile not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}